Create Single Page App Diagram
A sequence diagram depicting a user navigating to a single-page app. The browser sends a GET request to the server, and the server responds with an HTML document with script tags.
The browser executes javascript and fetches the notes from the server.
The server sends the JSON data notes. 
Then the browser renders the notes using the DOM-API